requirements.txt
streamlit==1.35.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.51.0
requests==2.31.0
